backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
langgraph==0.0.40
langchain>=0.1.1,<0.2.0
langchain-google-genai==0.0.6
langchain-community>=0.0.38
langsmith>=0.1.0,<0.2.0
pydantic>=2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
